\name{Factoshiny-package}
\alias{Factoshiny-package}
\alias{Factoshiny}
\docType{package}
\title{
Perform classical factorial analysis from FactoMineR within a Shiny app
~~ Factoshiny ~~
}
\description{
Factoshiny allows to perform CA, PCA, MFA, HCPC and MFA (classical functions from FactoMineR) within a Shiny app. The user can easily change the function parameters and the graphs parameters and can automatically see the restults of the change on the plot. All graphs can be downloaded in png, jpg and pdf.
}
\details{
\tabular{ll}{
Package: \tab Factoshiny\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-20\cr
License: \tab GPL (>= 2) \cr
}
Factoshiny have been created to be as easy to use as possible. Thus, only one parameter need to be informed in the several functions.
CAshiny, PCAshiny, MCAshiny and MFAshiny need a dataset to work. HCPCshiny need the result of a factorial analysis or a dataframe.
}
\author{
Pauline Vaissie, Astrid Monge, Francois Husson

Maintainer: Who to complain to <Magalie.Houee@agrocampus-ouest.fr>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
}
