\name{Factoshiny-package}
\alias{Factoshiny-package}
\alias{Factoshiny}
\docType{package}
\title{
Perform classical factorial analysis from FactoMineR within a Shiny app
~~ Factoshiny ~~
}
\description{
Factoshiny allows to perform CA, PCA, MFA, HCPC and MFA (classical functions from FactoMineR) within a Shiny app. The user can easily change the function parameters and the graphs parameters and can automatically see the restults of the change on the plot. All graphs can be downloaded in png, jpg and pdf.
}
\details{
Factoshiny has been created to be as easy to use as possible. Thus, only the dataset is needed to be informed in the main function.
}
\author{
Pauline Vaissie, Astrid Monge, Francois Husson

Maintainer: Who to complain to <Magalie.Houee@agrocampus-ouest.fr>
}
\references{
A website: \url{http://factominer.free.fr/graphs/factoshiny.html}

Some videos: \url{https://youtu.be/6y0NDDQcIXo}
}
\keyword{ package }
